Hides hidden connections on character page, 2 months endorsement expiry (#206)
    Endorsements expire after 2 months, instead of 3
    First stage Daeva addiction connections are now hidden on their preferences page, keep a little black book instead :)

+++ code/modules/vtr13/preferences/html_procs/connections_page.dm
@@ -4,6 +4,8 @@
 	dat += "<table width=width='70%' cellpadding='5' align='center'>"
 	character_connections = SScharacter_connection.get_character_connections(parent.ckey, src.real_name)
 	for(var/datum/character_connection/connection in character_connections)
+		if(connection.hidden)
+			continue
 		dat += "<tr><td><b>[connection.connection_desc]</b></td><td><a href='byond://?_src_=prefs;preference=connection;task=delete_connection;connection_id=[connection.group_id];'>Delete</a></td></tr>"
 	
 	dat += "</table>"
